/hardware/libhardware_legacy: Android SUPL NI client (Revision 2)
Network Initiated support for AGPS on UMTS (SUPL) is added to Android framework and the
GPS Hardware Abstraction Layer (HAL). It displays a notification message on the UI when
receiving a notification only NI request, or a dialog box when receiving a notification/verification
request. User responses are sent back to the HAL and GPS libraries (e.g. libloc_api).

CRs-Fixed: 186692

 On branch gps-chkin

Conflicts:

	include/hardware_legacy/gps_ni.h